requirements.txt
tree-sitter
firebase_admin
pyrebase4
pydantic